Refactored Cadence NFT List view
Closes #1164
