+++ apps/extension/src/ui/components/NFTs/EmptyStatus.tsx
