+++ apps/extension/src/ui/views/Setting/AccountList/stories/LinkedDetail.stories.tsx
@@ -2,7 +2,7 @@ import { type Meta, type StoryObj } from '@storybook/react-webpack5';
 import React from 'react';
 import { withRouter } from 'storybook-addon-remix-react-router';
 
-import type { NFTCollections } from '@onflow/flow-wallet-shared/types/nft-types';
+import type { CollectionNftList } from '@onflow/flow-wallet-shared/types/nft-types';
 import type { WalletAccount } from '@onflow/flow-wallet-shared/types/wallet-types';
 
 import {
@@ -52,7 +52,7 @@ const mockParentAccount = {
 };
 
 // Mock NFT collections data
-const mockNFTCollections: NFTCollections[] = [
+const mockNFTCollections: CollectionNftList[] = [
   {
     collection: {
       id: 'TestCollection',
